Fix para el problema del servicio escribiendo varias veces el cierre. Realmente es un problema de Async y como estoy ejecutando las tareas, pero esto lo soluciona por el momento.

+++ dbOPS.py
@@ -649,16 +649,28 @@ class DB:
 		self.conn.close()
 	def closeTrade(self, trade):
 		cur = self.tryConnect()
-		query = f"INSERT INTO traded (`openTime`, `symbol`, `entryStra`, `exitStra`, `qty`, `price`, `baseQty`, `closeTime`, `sellPrice`,`baseProfit`) VALUES ('{trade['openTime']}', '{trade['symbol']}', '{trade['entryStra']}', '{trade['exitStra']}','{trade['qty']}', '{trade['price']}', '{trade['baseQty']}', '{trade['closeTime']}', '{trade['sellPrice']}', '{trade['baseProfit']}');"
-		print(query)
+		query = f"SELECT COUNT(*) FROM traded WHERE openTime = '{trade['openTime']}' AND symbol = '{trade['symbol']}' AND qty = '{trade['qty']}'"
+		#print(query)
 		cur.execute(query)
-		self.conn.commit()
-		print("Insertando CIERRE")
-		query = f"DELETE FROM trading WHERE symbol = '{trade['symbol']}'"
-		cur.execute(query)
-		self.conn.commit()
-		self.conn.close()
-		print("Eliminando ABIERTO")
+		isClosed = False
+		for point in cur:
+			try:
+				isClosed =  bool(point[0])
+			except:
+				isClosed = False
+		if isClosed == False:
+			query = f"INSERT INTO traded (`openTime`, `symbol`, `entryStra`, `exitStra`, `qty`, `price`, `baseQty`, `closeTime`, `sellPrice`,`baseProfit`) VALUES ('{trade['openTime']}', '{trade['symbol']}', '{trade['entryStra']}', '{trade['exitStra']}','{trade['qty']}', '{trade['price']}', '{trade['baseQty']}', '{trade['closeTime']}', '{trade['sellPrice']}', '{trade['baseProfit']}');"
+			print(query)
+			cur.execute(query)
+			self.conn.commit()
+			print("Insertando CIERRE")
+			query = f"DELETE FROM trading WHERE symbol = '{trade['symbol']}'"
+			cur.execute(query)
+			self.conn.commit()
+			self.conn.close()
+			print("Eliminando ABIERTO")
+		else:
+			pass
 
 if __name__ == "__main__":
 	db1 = DB()
